requirements.txt
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
google-api-core==2.29.0
google-auth==2.48.0
google-cloud-pubsub==2.35.0
google-cloud-secret-manager==2.26.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.78.0
grpcio-status==1.78.0
idna==3.11
importlib-metadata==8.7.1
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1-modules==0.4.2
pycparser==3.0
requests==2.32.5
rsa==4.9.1
typing-extensions==4.15.0
urllib3==2.6.3
zipp==3.23.0
